+++ tests/InfluxDbClientTest.php
@@ -34,13 +34,27 @@ class InfluxDbClientTest extends TestCase
         parent::setUp();
     }
 
+    public function testConnectDatabaseOnGet()
+    {
+        $this->client = $this->getMock(
+            'Popstas\Transmission\Console\InfluxDbClient',
+            ['connectDatabase'],
+            [$this->influxDb, 'dbname']
+        );
+        $database = $this->getMock('InfluxDB\Database', [], ['dbname', $this->influxDb]);
+        $this->client->expects($this->once())->method('connectDatabase')->willReturn($database);
+
+
+        $this->client->writePoints([]);
+    }
+
     public function testConnectExistsDatabase()
     {
         $database = $this->getMock('InfluxDB\Database', [], ['dbname', $this->influxDb]);
         $database->method('exists')->will($this->returnValue(true));
         $this->influxDb->method('selectDB')->will($this->returnValue($database));
 
-        $this->client->connectDatabase();
+        $this->client->setDatabase($this->client->connectDatabase());
         $this->client->connectDatabase();
     }
 
@@ -72,6 +86,15 @@ class InfluxDbClientTest extends TestCase
 
     public function testBuildPoint()
     {
+        $database = $this->getDatabaseMockReturnsResultSet([]);
+        $this->client->setDatabase($database);
+
+        $torrent = $this->expectedTorrentList[0];
+        $this->client->buildPoint($torrent, 'localhost');
+    }
+
+    private function getDatabaseMockReturnsResultSet($value)
+    {
         $database = $this->getMock('InfluxDB\Database', [], ['dbname', $this->influxDb]);
 
         $queryBuilder = $this->getMock('InfluxDB\Query\Builder', ['getResultSet'], [$database]);
@@ -80,14 +103,11 @@ class InfluxDbClientTest extends TestCase
             ->setConstructorArgs([''])
             ->disableOriginalConstructor()
             ->getMock();
-        $resultSet->method('getPoints')->willReturn([]);
+        $resultSet->method('getPoints')->willReturn($value);
         $queryBuilder->method('getResultSet')->will($this->returnValue($resultSet));
         $database->method('getQueryBuilder')->willReturn($queryBuilder);
 
-        $this->client->setDatabase($database);
-
-        $torrent = $this->expectedTorrentList[0];
-        $this->client->buildPoint($torrent, 'localhost');
+        return $database;
     }
 
     public function testWritePoints()
@@ -97,5 +117,23 @@ class InfluxDbClientTest extends TestCase
         $this->client->setDatabase($database);
 
         $this->client->writePoints([]);
+    }
+
+
+    public function testGetTorrentSum()
+    {
+        $database = $this->getDatabaseMockReturnsResultSet([['uploaded_last' => 123]]);
+        $this->client->setDatabase($database);
+        $this->assertEquals(
+            123,
+            $this->client->getTorrentSum($this->expectedTorrentList[0], 'uploaded_last', 'localhost', 7)
+        );
+
+        $database = $this->getDatabaseMockReturnsResultSet([]);
+        $this->client->setDatabase($database);
+        $this->assertEquals(
+            0,
+            $this->client->getTorrentSum($this->expectedTorrentList[0], 'uploaded_last', 'localhost', 7)
+        );
     }
 }